refactor: move parsers to lib/parsers/
