+++ src/agent-transpiler-overrides.test.ts
@@ -4,7 +4,7 @@ import {
   copilotAgentTranspiler,
   claudeCodeAgentTranspiler,
 } from './agent-transpilers.ts';
-import { mergeOverrides } from './override-parser.ts';
+import { mergeOverrides } from './lib/parsers/index.ts';
 import type { CanonicalAgent, DiscoveredItem } from './types.ts';
 
 // ---------------------------------------------------------------------------
