+++ src/context-check.ts
@@ -1,7 +1,7 @@
 import pc from 'picocolors';
 import { resolve } from 'path';
 import { cloneRepo, cleanupTempDir } from './lib/git/index.ts';
-import { parseSource } from './source-parser.ts';
+import { parseSource } from './lib/parsers/index.ts';
 import { discover, filterByType } from './context-discovery.ts';
 import { executeInstallPipeline } from './context-installer.ts';
 import {
