+++ src/find.ts
@@ -3,7 +3,7 @@ import * as p from '@clack/prompts';
 import { runAdd } from './add.ts';
 import { parseAddOptions } from './add-options.ts';
 import { track } from './telemetry.ts';
-import { isRepoPrivate, parseOwnerRepo } from './source-parser.ts';
+import { isRepoPrivate, parseOwnerRepo } from './lib/parsers/index.ts';
 import { fetchRepoTree } from './lib/git/index.ts';
 import { discoverRemoteContext, type RemoteContextSummary } from './find-discovery.ts';
 import { RESET, BOLD, DIM, TEXT, CYAN, MAGENTA, YELLOW } from './utils.ts';
