+++ src/lib/parsers/agent-parser.ts
@@ -1,11 +1,11 @@
 import matter from 'gray-matter';
-import type { CanonicalAgent, AgentOverrideFields, TargetAgent } from './types.ts';
+import type { CanonicalAgent, AgentOverrideFields, TargetAgent } from '../../types.ts';
 import {
   SUPPORTED_SCHEMA_VERSION,
   validateName,
   validateDescription,
   validateSchemaVersion,
-} from './validation.ts';
+} from '../../validation.ts';
 import { extractOverrides } from './override-parser.ts';
 
 // ---------------------------------------------------------------------------
